+++ src/octuple.ts
@@ -1,3 +1,5 @@
+import { Avatar } from './components/Avatar';
+
 import { Badge } from './components/Badge';
 
 import {
@@ -30,6 +32,8 @@ import { Menu } from './components/Menu';
 
 import { Modal, ModalSize } from './components/Modal';
 
+import { Pagination, PaginationLayoutOptions } from './components/Pagination';
+
 import { Pill, PillSize, PillType } from './components/Pills';
 
 import {
@@ -44,6 +48,8 @@ import {
 import { InfoBar, InfoBarType } from './components/InfoBar';
 
 import { SnackbarContainer, Snackbar, snack } from './components/Snackbar';
+
+import { Spinner, SpinnerSize } from './components/Spinner';
 
 import { Tabs, Tab, TabVariant } from './components/Tabs';
 
@@ -64,6 +70,7 @@ import { useOnClickOutside } from './hooks/useOnClickOutside';
 import { useScrollLock } from './hooks/useScrollLock';
 
 export {
+    Avatar,
     Badge,
     ButtonIconAlign,
     ButtonShape,
@@ -89,6 +96,8 @@ export {
     Modal,
     ModalSize,
     NeutralButton,
+    Pagination,
+    PaginationLayoutOptions,
     Panel,
     PanelPlacement,
     PanelSize,
@@ -102,6 +111,8 @@ export {
     snack,
     Snackbar,
     SnackbarContainer,
+    Spinner,
+    SpinnerSize,
     Tabs,
     Tab,
     TabVariant,